improve HealthKit uploads, update deps, restructure some things, overall cleanup, add "upcoming events" tab
