+++ MyHeartCounts.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -204,8 +204,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/StanfordSpezi/SpeziAccount.git",
       "state" : {
-        "branch" : "lukas/swift-syntax-update-2nd-attempt",
-        "revision" : "8bbf042600928482e545343278ca5af6b52aa948"
+        "revision" : "9bb3d277f7f0e7f2e5abbe22040fd8fd626042f0",
+        "version" : "2.1.6"
       }
     },
     {
@@ -231,8 +231,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/StanfordSpezi/SpeziFoundation.git",
       "state" : {
-        "revision" : "f1506d2b4d4d23e9634668c741cd1cb985aad111",
-        "version" : "2.1.3"
+        "revision" : "342e463af6694bc279a8c0aa24791f81254dfe6c",
+        "version" : "2.1.4"
       }
     },
     {
@@ -258,8 +258,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/StanfordSpezi/SpeziNotifications.git",
       "state" : {
-        "revision" : "21427fff437e6bc17db74b125fa43970a2ab9b0b",
-        "version" : "1.0.5"
+        "revision" : "4a9ff11f41a0d3d3cf70e4696dbf604afe4330e6",
+        "version" : "1.0.6"
       }
     },
     {
@@ -285,8 +285,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/StanfordSpezi/SpeziScheduler.git",
       "state" : {
-        "branch" : "lukas/swift-syntax-update-2nd-attempt",
-        "revision" : "ad6c93364173c38b7a91f5e1bbfb3da61c359431"
+        "revision" : "29aa1b80835cbb00acf904a762b9e9f06a4949ce",
+        "version" : "1.2.5"
       }
     },
     {
@@ -303,8 +303,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/StanfordSpezi/SpeziStudy.git",
       "state" : {
-        "revision" : "e20a81de83f511d359c108231de3752a838054ad",
-        "version" : "0.1.0"
+        "revision" : "badd0563462ac1983ed75f06649e7e9ab1e73c1e",
+        "version" : "0.1.1"
       }
     },
     {
