+++ MyHeartCounts.xcodeproj/xcshareddata/xcschemes/MyHeartCounts.xcscheme
@@ -83,7 +83,7 @@
          </CommandLineArgument>
          <CommandLineArgument
             argument = "--overrideFirebaseConfigOnDevice"
-            isEnabled = "NO">
+            isEnabled = "YES">
          </CommandLineArgument>
          <CommandLineArgument
             argument = "--showOnboarding"
